Add files via upload
HairStory is tag name for the project . It is the application of booking time slot of hair salon. Generally peoples waiting hours at the salon , it is very time consuming so this application will help the people for doing his/her salon work without wasting a time . Generally I made the front end structure for that application using Html ,Html5, Css, Css3 and framework like bootstrap.